Refactor the python codes part
Authored-by: Tingfang Bao <bbao@pivotal.io>
